pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "walter-relance-test"
version = "0.1.0"
description = "Test technique senior - moteur de relance apprenants"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.5",
  "httpx>=0.27",
  "pyyaml>=6",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "respx>=0.21",
  "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
